Fold s3_drop helpers into the Roseville adapter per review
